+++ src/Http/WikidataLocationAssignmentAction.php
@@ -11,6 +11,7 @@ use Fisharebest\Webtrees\Http\Exceptions\HttpBadRequestException;
 use Fisharebest\Webtrees\Registry;
 use Fisharebest\Webtrees\Validator;
 use Hartenthaler\Webtrees\Module\ExternalPlacesModule\Domain\WikidataIdentifier;
+use Hartenthaler\Webtrees\Module\ExternalPlacesModule\External\ExternalPerson;
 use Hartenthaler\Webtrees\Module\ExternalPlacesModule\External\ExternalProviderRegistry;
 use Hartenthaler\Webtrees\Module\ExternalPlacesModule\Gedcom\WikidataLocationAssignmentService;
 use Hartenthaler\Webtrees\Module\ExternalPlacesModule\Geo\Coordinates;
@@ -92,6 +93,53 @@ final class WikidataLocationAssignmentAction implements RequestHandlerInterface
             }
             $added = $service->addAddress($location, $address);
             FlashMessages::addMessage($added ? I18N::translate('The address has been added to the shared place.') : I18N::translate('The address was not added because it already exists or is empty.'), $added ? 'success' : 'danger');
+        } elseif ($operation === 'add-person') {
+            $providerKey = Validator::parsedBody($request)->string('person_provider', '');
+            $provider = (new ExternalProviderRegistry())->byKey($providerKey);
+            $externalId = Validator::parsedBody($request)->string('person_id', '');
+            $identifier = $provider?->identifier($externalId);
+            $label = trim(Validator::parsedBody($request)->string('person_label', ''));
+            $birthDate = trim(Validator::parsedBody($request)->string('person_birth', ''));
+            $deathDate = trim(Validator::parsedBody($request)->string('person_death', ''));
+            $sex = strtoupper(trim(Validator::parsedBody($request)->string('person_sex', '')));
+            $wikidataId = strtoupper(trim(Validator::parsedBody($request)->string('person_wikidata', '')));
+            $factgridId = strtoupper(trim(Validator::parsedBody($request)->string('person_factgrid', '')));
+            $wikiTreeId = trim(Validator::parsedBody($request)->string('person_wikitree', ''));
+            $relationship = Validator::parsedBody($request)->string('person_relationship', '');
+            $from = trim(Validator::parsedBody($request)->string('person_from', ''));
+            $until = trim(Validator::parsedBody($request)->string('person_until', ''));
+            if ($identifier === null || !in_array($providerKey, ['wikidata', 'factgrid'], true)) {
+                throw new HttpBadRequestException(I18N::translate('The external provider identifier is invalid.'));
+            }
+            if ($label === '' || mb_strlen($label) > 240 || preg_match('/[\r\n]/', $label) === 1) {
+                throw new HttpBadRequestException(I18N::translate('The external person has no usable name.'));
+            }
+            if (!in_array($relationship, ['Owner', 'Occupant'], true)) {
+                throw new HttpBadRequestException(I18N::translate('The external-person relationship is invalid.'));
+            }
+            if ($sex !== '' && !in_array($sex, ['M', 'F', 'U', 'X'], true)) {
+                throw new HttpBadRequestException(I18N::translate('The external person has an invalid sex value.'));
+            }
+            $externalLinks = [];
+            if ($wikidataId !== '' && preg_match('/^Q[1-9][0-9]*$/', $wikidataId) === 1) {
+                $externalLinks['Wikidata'] = 'https://www.wikidata.org/entity/' . $wikidataId;
+            }
+            if ($factgridId !== '' && preg_match('/^Q[1-9][0-9]*$/', $factgridId) === 1) {
+                $externalLinks['Factgrid'] = 'https://database.factgrid.de/entity/' . $factgridId;
+            }
+            if ($wikiTreeId !== '' && preg_match('/^[\p{L}][\p{L}\p{M}0-9._-]{0,119}$/u', $wikiTreeId) === 1) {
+                $externalLinks['WikiTree'] = 'https://www.wikitree.com/wiki/' . rawurlencode($wikiTreeId);
+            }
+            $person = new ExternalPerson($providerKey, $externalId, $identifier->url, $label, $birthDate !== '' ? $birthDate : null, $deathDate !== '' ? $deathDate : null, $externalLinks, $sex !== '' ? $sex : null);
+            $result = $service->addPerson($location, $person, $relationship, $from !== '' ? $from : null, $until !== '' ? $until : null);
+            $message = match ($result) {
+                'added' => I18N::translate('The external person has been added to the shared place.'),
+                'already-associated' => I18N::translate('This external person is already linked to the shared place.'),
+                'invalid-provider' => I18N::translate('The external person was not added because the provider identifier is invalid.'),
+                'not-authorized' => I18N::translate('You are not authorized to add a person to this shared place.'),
+                default => I18N::translate('The individual was created, but it could not be linked to the shared place.'),
+            };
+            FlashMessages::addMessage($message, $result === 'added' ? 'success' : 'danger');
         } else {
             throw new HttpBadRequestException(I18N::translate('Invalid Wikidata assignment operation.'));
         }